impl

+++ src/main/java/com/tsoft/civilization/combat/CombatDamage.java
@@ -11,7 +11,7 @@ public class CombatDamage {
 
     public static final CombatDamage ZERO = CombatDamage.builder().damage(0).build();
 
-    private final int damage;
+    private final double damage;
 
     public CombatDamage add(CombatDamage other) {
         return CombatDamage.builder()
