+++ src/main/java/com/tsoft/civilization/combat/CombatStrength.java
@@ -11,21 +11,21 @@ public class CombatStrength {
 
     // Ranged attack
     private final int rangedAttackLevel;      // Unit's level in ranged attacks
-    private final int rangedAttackStrength;   // Unit's ranged attack strength
+    private final double rangedAttackStrength;   // Unit's ranged attack strength
     private final int rangedAttackRadius;     // Ranged attack radius
-    private final int rangedAttackExperience; // Experience during an attack
-    private final int rangedBackFireStrength; // Backfire strength
+    private final double rangedAttackExperience; // Experience during an attack
+    private final double rangedBackFireStrength; // Backfire strength
 
     // Melee attack
     private final int meleeAttackLevel;       // Unit's level in melee attacks
-    private final int meleeAttackStrength;    // Unit's ranged attack strength
-    private final int meleeAttackExperience;  // Experience during an attack
-    private final int meleeBackFireStrength;  // Backfire strength
+    private final double meleeAttackStrength;    // Unit's ranged attack strength
+    private final double meleeAttackExperience;  // Experience during an attack
+    private final double meleeBackFireStrength;  // Backfire strength
 
     // Defense
     private final int defenseLevel;           // Unit's defense level
-    private final int defenseStrength;        // Unit's defense strength
-    private final int defenseExperience;      // Experience during a defense
+    private final double defenseStrength;        // Unit's defense strength
+    private final double defenseExperience;      // Experience during a defense
 
     // Options
     private final boolean isDestroyed;        // Was destroyed during a step
